requirements.txt
Flask
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
numpy
